requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
